Parse source from files and save it in the DB
So far the code was parsing only the IP info, set it in a 1d array,
manipulated the data (in order to get distinct values and make the
data set smaller), and saved the data in the IP column.

This change adds also the source information. The process is this:
1. Get the value of the 'Maintainer' line using regexp
2. Save the data in an array of JSON objects (ip, source)
3. Group the array objects by IP (function: groupBy)
4. Save the data in a 2D array by adding the IP and concatenating
all the sources (function: mergeIPSources)
5. Update the syncData function of queries.js in order to process
the 2D array and save both IP and Source(s) in the DB.